+++ packages/atomic/cypress/fixtures/test-fixture.ts
@@ -183,6 +183,7 @@ export class TestFixture {
           {
             method: 'POST',
             url: '**/rest/search/v2?*',
+            times: 9999,
           },
           (request) => {
             request.reply((response) => {